pyproject.toml
[project]
name = "tesseract-api"
version = "0.1.0"
description = "Simple FastAPI service that solves captchas using Tesseract OCR"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.117.1",
    "pytesseract>=0.3.13",
    "pillow>=10.4.0",
    "uvicorn>=0.37.0",
    "python-multipart>=0.0.20",
]
